+++ examples/01_quote_of_the_day/lib/quote_of_the_day.ex
@@ -1,6 +1,6 @@
 defmodule QuoteOfTheDay do
   use Application
-  @behaviour Ace.TCP.Server
+  use Ace.Application
 
   def start(_type, _args) do
     import Supervisor.Spec, warn: false